+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Updated Exceptional call signature
+
 ## v0.2.3 (2024-08-21)
 * Converted consts to protected PascalCase
 * Made PHP8.1 minimum version